feature(Settings): Refactoring code, splitting out fields module
